Introduce BlogPost writers
### What is a Writer for?

A writer class helps to write a blog post. It will help in things like
extracting a list of topics for a message or giving it a title. This
helps to keep the BlogPost class cleaner by delegating the work to
other objects.

### The Writer interface

A Writer is a class that follows the following interface:

1. `Writer#extract_topics_from(message_text)`
2. `Writer#give_title_to(message)`
3. `Writer#summarize(message_text)`

Instead of inheritance and NotImplementedYet errors, I went with shared examples
to ensure the required methods were implemented. [This blog post] talks
more about this approach.

### SimpleWriter

The SimpleWriter class is basically a wrapper around the current
behavior. It does the most simple things:

- Returns the permalink as the message title
- Returns the message texts as a summary, i.e., it does no summarization
at all.
- Delegates the topic extraction to the TopicExtractor module

While this might look silly, it was created to pave the way
for other future writers. For example, I'm planning to add an `OpenAIWriter` in
a follow-up PR. If folks want, they can use that, otherwise
GoldMiner will fall back to the SimpleWriter.

[This blog post]: https://shiroyasha.io/interfaces-in-ruby.html