Introduce OpenAIWriter
This writer uses the [OpenAI API] to extract topics, generate titles and
summarize messages. It has a fallback writer to use in cases where the API might
return an error.

Some prompts had to be specific to get good results. We might have to tweak them
a bit if we feel the results could be more satisfactory.

Here's [an example] of a blog post generated by this writer.

[OpenAI API]: https://openai.com/api/
[an example]: https://gist.github.com/MatheusRich/6bed395fcb525327d46faf3b79a95d87