Fetch Slack message author name
This removes some manual work. Before, it was up to the
blog post author to get the message author names based on their
Slack usernames.

Unfortunately it requires another request to the Slack API (names aren't
available on the message search response), which creates an N+1
problem. While it's bad for performance, we probably won't have more
than 10 messages, so it's not that many extra requests.

I also introduced the Slack::Message class. It doesn't have any
particular behavior, but I already have plans for extra features.
It implements #[], so we can use instances like hashes. This makes it
easier to swap hashes for this class without much refactoring.
For instance, I didn't touch any blog post writer class with this
change.

Fixes #6