Add author links on article
Up to now, the article writer had to manually add links for every author
mentioned in the article. This PR automates of this process.

Now we load an `AuthorConfig` that has a preferred link for each author. If the
author is not in the config, we use the default `#to-do` link.

This class also has two additional advantages:

1. It allows future improvements, like adding a preferred name, instead of
   automatically using the one from Slack.
2. It allows users to update their own links, instead of asking someone else to
   do it. This is in line with [our value of self-management].

Other changes:

- I updated the prompt on `OpenAiWriter`, to prevent it from removing the author
links.
- I update some specs to use a Factory instead of ad-hoc objects.

[our value of self-management]: https://thoughtbot.com/playbook/our-company/purpose-mission-values#self-management