Rewrite Slack::Client
This class used to know too much about the app-specific needs.
It's now a thin wrapper around the Slack API client gem.

A major change though, is that it automatically fetches the message
author real name, so it issues an additional API call for each message.
To remedy the performance hits, all the author names are requested in
parallel, with a cache to avoid unnecessary requests.