Introduce `GoldMiner::SlackExplorer`
This class is responsible for "exploring" Slack and "mining" interesting
messages. Those messages, later become a blog post.

Now that we have a simple Slack client, we can put the business logic
in the explorer. This also opens the door for more explorers in the future,
like a Twitter explorer.