Introduce Gold Miner distributors
Distribution is the last step after mining and smithing gold. It's the final step,
where we deliver something useful to the world. We do that using distributor
objects.

Distributors are objects that respond to `#distribute` and perform some
kind of "delivery". For now, we have the `TerminalDistributor` which prints the
generated blog post to the terminal. I'm [planning to add] a GitHub distributor
that will open a pull request for the generated blog post.

Unlike Smiths, that take GoldContainers with GoldNuggets, I haven't found
a good abstraction for what Distributors accept. We only produce blog posts,
so I'm using that for now. It's not generic, but I'll wait until we have more
examples of distributors to find a good abstraction.

[planning to add]: https://github.com/thoughtbot/gold_miner/issues/1