Rewrite `GoldMiner` main entrypoint
Previously, the `GoldMiner` module (now a class) had methods for each part of
the mining process: exploration, smithing, and distribution.

The orchestration of when and how these methods were called was done in the
`exe/gold_miner` script. This wasn't ideal since there's no way to guarantee the
order in which the methods are called, and a script is arguably not the best
place to put this logic. Furthermore, that arrangement wasn't flexible, as there
was no way to specify different explorers, smiths, or distributors.

This commit rewrites the `GoldMiner` class to have an `initialize` method that
receives the explorer, smith, and distributor as arguments. It also adds a
`#mine` method that orchestrates the mining process.